+++ NEWS.md
@@ -2,6 +2,8 @@ ggptt 0.3.2
 ============
 
 * `set_ptt()` to set theme_ptt and ptt scales.
+* `ggptt_pal()` for all color palettes in the package and a new palette "vnk".
+  
 
 ggptt 0.3.1
 ============